requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# AI/ML
google-generativeai==0.3.2

# Environment variables
python-dotenv==1.0.0

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

